backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# LangChain + integrations  (upgraded for Python 3.12 compatibility)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-text-splitters>=0.3.0
langsmith>=0.1.147

# Embeddings
sentence-transformers==2.7.0
huggingface-hub>=0.23.1

# Vector store
chromadb==0.5.0

# Document loaders
pypdf==4.2.0
python-docx==1.1.2

# Database
sqlalchemy==2.0.30
aiosqlite==0.20.0          # for async SQLite

# Security & utilities
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testing
pytest==8.2.0
httpx==0.27.0              # for TestClient
